[REF] sale_commission: Split common setup for tests
